[REFACTOR] Friendship Entity UserService code move to Friendship Dict
