+++ src/main/java/com/PuzzleU/Server/user/service/UserService.java
@@ -377,58 +377,7 @@ public class UserService {
         return ResponseUtils.ok(SuccessResponse.of(HttpStatus.OK, "프로필 기본 정보 수정 완료"), null);
     }
 
-    // 특정 유저에게 친구신청을 거는 것
-    public ApiResponseDto<SuccessResponse> registerFriend(UserDetails loginUser, Long userId) {
-        User user1 = userRepository.findByUsername(loginUser.getUsername())
-                .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_USER));
-        // friendship이 만들어져야한다
-        // 무조건 거는 사람이 user1, 받는 사람이 user2로 되도록
-        Optional<User> userOptional = userRepository.findById(userId);
-        User user2 = userOptional.orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_USER));
-        // 만약에 user1과 2가 존재한다면 끝
-        Optional<FriendShip> friendShip_check = friendshipRepository.findByUser1AndUser2(user1, user2);
 
-        if (friendShip_check.isPresent()) {
-            return ResponseUtils.ok(SuccessResponse.of(HttpStatus.OK, "이미 진행된 친구신청입니다."), null);
-        } else {
-            // 친구 관계가 존재하지 않는 경우, 새로운 친구 관계 생성
-            FriendShip friendShip = new FriendShip();
-            friendShip.setUser2(user2);
-            friendShip.setUser1(user1);
-            friendShip.setUserStatus(false);
-            friendshipRepository.save(friendShip);
-            return ResponseUtils.ok(SuccessResponse.of(HttpStatus.OK, "친구신청이 완료되었습니다"), null);
-        }
-
-    }
-    public ApiResponseDto<SuccessResponse> responseFriend(UserDetails loginUser, Long userId) {
-        System.out.println(loginUser);
-        System.out.println(userId);
-        User user1 = userRepository.findByUsername(loginUser.getUsername())
-                .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_USER));
-        System.out.println(user1);
-        Optional<User> userOptional = userRepository.findById(userId);
-        User user2 = userOptional.orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_USER));
-        System.out.println(user2);
-        Optional<FriendShip> friendShip_check = friendshipRepository.findByUser1AndUser2(user1, user2);
-        FriendShip friendShip = friendShip_check.orElseThrow(()-> new RestApiException(ErrorType.NOT_FOUND_FRIENDSHIP));
-        System.out.println(friendShip);
-        friendShip.setUserStatus(true);
-        friendshipRepository.save(friendShip);
-        return ResponseUtils.ok(SuccessResponse.of(HttpStatus.OK, "친구수락이 완료되었습니다"), null);
-
-    }
-    public ApiResponseDto<SuccessResponse> deleteFriend(UserDetails loginUser, Long userId) {
-        User user1 = userRepository.findByUsername(loginUser.getUsername())
-                .orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_USER));
-        Optional<User> userOptional = userRepository.findById(userId);
-        User user2 = userOptional.orElseThrow(() -> new RestApiException(ErrorType.NOT_FOUND_USER));
-        Optional<FriendShip> friendShip_check = friendshipRepository.findByUser1AndUser2(user1, user2);
-        FriendShip friendShip = friendShip_check.orElseThrow(()-> new RestApiException(ErrorType.NOT_FOUND_FRIENDSHIP));
-        friendshipRepository.delete(friendShip);
-        return ResponseUtils.ok(SuccessResponse.of(HttpStatus.OK, "친구신청이 완료되었습니다"), null);
-
-    }
